requirements.txt
chromadb==1.0.16
groq==0.30.0
python-dotenv==1.1.0
Requests==2.32.3
sentence_transformers==5.0.0
streamlit==1.45.0
pysqlite3-binary==0.5.4